+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Modules/About/Internal/LitsItems/AccessibilityStatement/AccessibilityStatement.swift
@@ -20,6 +20,7 @@
 //
 //
 
+import DeclarationAccessibility
 import SwiftUI
 
 struct AboutAccessibilityStatementItemConfig: ODSAboutListItemConfig {
@@ -41,26 +42,10 @@ struct AboutAccessibilityStatementItemConfig: ODSAboutListItemConfig {
         title = "Accessibility Statement"
         icon = Image("ic_accessibility", bundle: Bundle.ods)
         priority = .accessibilityStatement
-        target = .destination(AnyView(AccessibilityStatement(statementConfig: statementConfig)))
-    }
-}
-
-private struct AccessibilityStatement: View {
-
-    // =======================
-    // MARK: Stored Properties
-    // =======================
-
-    let statementConfig: ODSAboutAccessibilityStatement
-
-    // ==========
-    // MARK: Body
-    // ==========
-
-    var body: some View {
-        ScrollView {
-            Text("The Accessibility statement can be generated using the Orange's V11ydette: https://la-va11ydette.orange.com/")
-                .padding(ODSSpacing.m)
-        }
+        target = .destination(AnyView(
+            DeclarationView(xmlFileName: statementConfig.fileName,
+                            selectedTheme: .orange,
+                            url: statementConfig.reportDetail.absoluteString)
+            ))
     }
 }