+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Screens/Modules/CardListAndCollection/GrifOfSmallCards.swift
@@ -74,4 +74,3 @@ struct GrifOfSmallCard_Previews: PreviewProvider {
     }
 }
 #endif
-